Move Mesh and MeshRenderable into own files
