+++ FloodForge/src/world/room/Room.cs
@@ -1,4 +1,5 @@
 using FloodForge.Popups;
+using FloodForge.Rendering;
 using Stride.Core.Extensions;
 
 namespace FloodForge.World;
@@ -919,182 +920,6 @@ public class Room : WorldDraggable {
 	}
 
 	#region Rendering
-
-	[StructLayout(LayoutKind.Sequential)]
-	protected class Mesh() {
-		public uint currentIndex = 0;
-		public List<Vertex> vertices = [];
-		public List<uint> indices = [];
-
-		public void Clear() {
-			this.vertices.Clear();
-			this.indices.Clear();
-			this.currentIndex = 0;
-		}
-
-		public void AddQuad(float xPos, float yPos, Themes.ThemeColor theme) {
-			this.AddQuad(new Vector2(xPos, yPos), Vector2.One, theme);
-		}
-
-		public void AddQuad(Vector2 centerPosition, Themes.ThemeColor theme) {
-			this.AddQuad(centerPosition, Vector2.One, theme);
-		}
-
-		public void AddQuad(float xPos, float yPos, float scale, Themes.ThemeColor theme) {
-			this.AddQuad(new Vector2(xPos, yPos), Vector2.One * scale, theme);
-		}
-
-		public void AddQuad(Vector2 centerPosition, float scale, Themes.ThemeColor theme) {
-			this.AddQuad(centerPosition, Vector2.One * scale, theme);
-		}
-
-		public void AddQuad(float xPos, float yPos, Vector2 scale, Themes.ThemeColor theme) {
-			this.AddQuad(new(xPos, yPos), scale, theme);
-		}
-
-		public void AddQuad(Vector2 centerPosition, Vector2 scale, Themes.ThemeColor theme) {
-			float x0 = centerPosition.x - (scale.x / 2);
-			float x1 = centerPosition.x + (scale.x / 2);
-			float y0 = centerPosition.y + (scale.y / 2);
-			float y1 = centerPosition.y - (scale.y / 2);
-			this.AddQuad(
-				new Vertex(x0, y0, theme),
-				new Vertex(x1, y0, theme),
-				new Vertex(x1, y1, theme),
-				new Vertex(x0, y1, theme)
-			);
-		}
-
-		public void AddQuad(Vertex a, Vertex b, Vertex c, Vertex d) {
-			this.vertices.Add(a);
-			this.vertices.Add(b);
-			this.vertices.Add(c);
-			this.vertices.Add(d);
-			this.indices.Add(this.currentIndex + 0);
-			this.indices.Add(this.currentIndex + 1);
-			this.indices.Add(this.currentIndex + 2);
-			this.indices.Add(this.currentIndex + 2);
-			this.indices.Add(this.currentIndex + 3);
-			this.indices.Add(this.currentIndex + 0);
-			this.currentIndex += 4;
-		}
-
-		public void AddTriangle(Vertex a, Vertex b, Vertex c) {
-			this.vertices.Add(a);
-			this.vertices.Add(b);
-			this.vertices.Add(c);
-			this.indices.Add(this.currentIndex + 0);
-			this.indices.Add(this.currentIndex + 1);
-			this.indices.Add(this.currentIndex + 2);
-			this.currentIndex += 3;
-		}
-	}
-
-	protected class MeshRenderable {
-		public Mesh mesh;
-		protected uint _vao = 0;
-		protected uint _vbo = 0;
-		protected uint _ebo = 0;
-		protected Dictionary<string, int> shaderVariableLocations = [];
-		protected Shader shaderToUse;
-
-		public unsafe MeshRenderable(Mesh mesh, Shader shaderToUse, VertexAttributeInformation[]? vertexAttributeInformation = null, string[]? shaderVariableLocations = null) {
-			this.mesh = mesh;
-			if (this._vao == 0) {
-				this._vao = Program.gl.GenVertexArray();
-				this._vbo = Program.gl.GenBuffer();
-				this._ebo = Program.gl.GenBuffer();
-			}
-
-			Span<Vertex> vertices = CollectionsMarshal.AsSpan(this.mesh.vertices);
-			Span<uint> indices = CollectionsMarshal.AsSpan(this.mesh.indices);
-
-			Program.gl.BindVertexArray(this._vao);
-
-			Program.gl.BindBuffer(BufferTargetARB.ArrayBuffer, this._vbo);
-			fixed (Vertex* ptr = vertices) {
-				Program.gl.BufferData(BufferTargetARB.ArrayBuffer, (nuint) (vertices.Length * sizeof(Vertex)), ptr, BufferUsageARB.StaticDraw);
-			}
-
-			Program.gl.BindBuffer(BufferTargetARB.ElementArrayBuffer, this._ebo);
-			fixed (uint* ptr = indices) {
-				Program.gl.BufferData(BufferTargetARB.ElementArrayBuffer, (nuint) (indices.Length * sizeof(uint)), ptr, BufferUsageARB.StaticDraw);
-			}
-			
-			if(vertexAttributeInformation != null) {
-				foreach (VertexAttributeInformation item in vertexAttributeInformation) {
-					Program.gl.VertexAttribPointer(item.index, item.size, item.type, item.normalised, item.stride, item.pointer);
-					Program.gl.EnableVertexAttribArray(item.index);
-				}
-			}
-			
-			this.shaderToUse = shaderToUse;
-			Program.gl.UseProgram(shaderToUse);
-			if(shaderVariableLocations != null) {
-				foreach (string name in shaderVariableLocations) {
-					this.shaderVariableLocations.Add(name, Program.gl.GetUniformLocation(shaderToUse, name));
-				}
-			}
-			Program.gl.UseProgram(0);
-
-			Program.gl.BindVertexArray(0);
-			Program.gl.BindBuffer(BufferTargetARB.ArrayBuffer, 0);
-			Program.gl.BindBuffer(BufferTargetARB.ElementArrayBuffer, 0);
-		}
-
-		public void UniformMatrix4(string name, bool transpose, ReadOnlySpan<float> value)
-			=> Program.gl.UniformMatrix4(this.shaderVariableLocations[name], transpose, value);
-		public void Uniform4(string name, float x, float y, float z, float w)
-			=> Program.gl.Uniform4(this.shaderVariableLocations[name], x, y, z, w);
-		public void Uniform1(string name, float val)
-			=> Program.gl.Uniform1(this.shaderVariableLocations[name], val);
-
-		public void PreDraw() {
-			Program.gl.BindVertexArray(this._vao);
-			Program.gl.UseProgram(Preload.RoomShader);
-		}
-
-		public unsafe void DoDraw() {
-			Program.gl.DrawElements(PrimitiveType.Triangles, (uint) this.mesh.indices.Count, DrawElementsType.UnsignedInt, (void*) 0);
-
-			Program.gl.BindVertexArray(0);
-			Program.gl.UseProgram(0);
-		}
-
-		public unsafe struct VertexAttributeInformation (uint index, int size, VertexAttribPointerType type, bool normalised, uint stride, void* pointer){
-			public uint index = index;
-			public int size = size;
-			public VertexAttribPointerType type = type;
-			public bool normalised = normalised;
-			public uint stride = stride;
-			public void* pointer = pointer;
-		}
-	}
-
-	protected readonly struct Vertex {
-		public readonly float x, y;
-		public readonly float r, g, b, a;
-
-		public Vertex(float x, float y, Color color) {
-			this.x = x;
-			this.y = y;
-			this.r = color.r;
-			this.g = color.g;
-			this.b = color.b;
-			this.a = color.a;
-		}
-
-		public Vertex(float x, float y, float r, float g, float b, float a = 1f) {
-			this.x = x;
-			this.y = y;
-			this.r = r;
-			this.g = g;
-			this.b = b;
-			this.a = a;
-		}
-	}
-
-	// IDEA - add MeshRenderReference object containing setup code for easier modifications
 	protected Mesh waterMesh = new Mesh();
 	protected MeshRenderable? waterRenderable;
 	protected Mesh roomMesh = new Mesh();
